# vulkan: optimize coopmat2 iq2/iq3 callbacks